requirements.txt
gunicorn
psycopg2-binary
django-cors-headers
djangorestframework
djangorestframework-simplejwt
whitenoise